+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Collected authentication method for telemetry in end process of cmdlet.
 * Fixed secrets exposure in example documentation.
 * Updated `Connect-AzAccount` to fix a display issue in PowerShell ISE [#24556].
 * Updated the reference of Azure PowerShell Common to 1.3.100-preview.